setdepartmentforguest
